Consolidate movie task settings into template.
Also modify secretfile.yml root directories to not including leading '/'.
This is instead moved to the config like:
```yml
move:
  to: '/{{ secrets.roots.media }}/{{ secrets.directories.tv_shows }}/{{ series_name }}/Season {{series_season}}/
  filename: '{{ series_name }} - {{ series_season|pad(2) }}x{{ series_episode|pad(2) }}'
```

This eliminates the validation error that occurs when a move path
starts with jinja replacement syntax and allows all directory paths to
be stored in and used from the secrettile instead of being hard-coded.